remoção de uma caminho a menos, conflito dos readme.md resolvido
